+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# next
+- now `onSkip` return a bool. if onSkip return false, the overlay will not be dismissed
+
 # 1.2.9
 - set `ignoringSemantics` false in `skip` button. Reolve [#156](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/156)
 - adds `ImageFilter`. Now is possible apply blur in the shadow. [#151](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/151)